New upstream release 6.0.7

+++ src/plugins/uploader.jumploader/class.JumploaderProcessor.php
@@ -375,10 +375,12 @@ class JumploaderProcessor extends AJXP_Plugin
                             $currentFile = $partitions[$i];
                             $currentFileName = $currentFile["tmp_name"];
                             $part = fopen($currentFileName, "r");
-                            while (!feof($part)) {
-                                $length += fwrite($newDest, fread($part, 4096));
+                            if(is_resource($part)){
+                                while (!feof($part)) {
+                                    $length += fwrite($newDest, fread($part, 4096));
+                                }
+                                fclose($part);
                             }
-                            fclose($part);
                             unlink($currentFileName);
                         }
                         $newFile["type"] = $partitions[0]["type"];
@@ -389,15 +391,18 @@ class JumploaderProcessor extends AJXP_Plugin
                         $newFile["destination"] = $partitions[0]["destination"];
                         $newPartitions[] = $newFile;
                     } else {
-                        $newDest = fopen($destStreamURL.$httpVars["partitionRealName"], "w");
+                        $current = $destStreamURL.$httpVars["partitionRealName"];
+                        $newDest = fopen($current, "w");
                         $fileHash = md5($httpVars["partitionRealName"]);
 
                         for ($i = 0; $i < $httpVars["partitionCount"] ; $i++) {
                             $part = fopen($destStreamURL."$fileHash.$fileId.$i", "r");
-                            while (!feof($part)) {
-                                fwrite($newDest, fread($part, 4096));
+                            if(is_resource($part)){
+                                while (!feof($part)) {
+                                    fwrite($newDest, fread($part, 4096));
+                                }
+                                fclose($part);
                             }
-                            fclose($part);
                             unlink($destStreamURL."$fileHash.$fileId.$i");
                         }
 
@@ -406,14 +411,17 @@ class JumploaderProcessor extends AJXP_Plugin
                 }
 
                 if (!self::$remote && (!self::$wrapperIsRemote || $relPath != $httpVars["partitionRealName"])) {
-                    $err = copy($current, $target);
+                    if($current != $target){
+                        $err = copy($current, $target);
+                    }
+                    else $err = true;
                 } else {
                     for ($i=0, $count=count($newPartitions); $i<$count; $i++) {
                         $driver->storeFileToCopy($newPartitions[$i]);
                     }
                 }
 
-                if ($err !== false) {
+                if ($current != $target && $err !== false) {
                     if(!self::$remote) unlink($current);
                     AJXP_Controller::applyHook("node.change", array(null, new AJXP_Node($target), false));
                 } else if ($current == $target) {
